tests/test_image_loader.py
"""Tests for ImageLoader module - handles loading and validating images."""
import pytest
from pathlib import Path
from PIL import Image

# Module will be: src.stitchify.image_loader.ImageLoader


class TestImageLoader:
    """Test suite for ImageLoader class."""
    
    def test_load_valid_png(self, simple_rgba_image):
        """Test loading a valid PNG image."""
        # Will implement: loader = ImageLoader(simple_rgba_image)
        # assert loader.image is not None
        # assert loader.width == 3
        # assert loader.height == 3
        pass
    
    def test_load_valid_jpg(self, test_images_dir):
        """Test loading a valid JPG image and converting to RGBA."""
        # JPG doesn't support transparency, should convert properly
        pass
    
    def test_load_invalid_path(self):
        """Test loading from non-existent path raises appropriate error."""
        # Should raise FileNotFoundError or similar
        pass
    
    def test_load_non_image_file(self, tmp_path):
        """Test loading a non-image file raises appropriate error."""
        text_file = tmp_path / "test.txt"
        text_file.write_text("Not an image")
        # Should raise appropriate PIL exception
        pass
    
    def test_get_dimensions(self, simple_rgba_image):
        """Test getting image dimensions."""
        # loader = ImageLoader(simple_rgba_image)
        # assert loader.get_dimensions() == (3, 3)
        pass
    
    def test_get_pixel_at_position(self, simple_rgba_image):
        """Test getting pixel data at specific coordinates."""
        # loader = ImageLoader(simple_rgba_image)
        # Red pixel at (0, 0)
        # assert loader.get_pixel(0, 0) == (255, 0, 0, 255)
        pass
    
    def test_get_pixel_out_of_bounds(self, simple_rgba_image):
        """Test getting pixel outside image bounds raises error."""
        # Should raise IndexError or similar
        pass
    
    def test_image_mode_conversion(self, test_images_dir):
        """Test that images are converted to RGBA mode."""
        # Create RGB image (no alpha)
        img = Image.new("RGB", (2, 2), (255, 0, 0))
        img_path = test_images_dir / "rgb_test.png"
        img.save(img_path)
        
        # loader = ImageLoader(img_path)
        # assert loader.image.mode == "RGBA"
        pass
    
    def test_load_with_pathlib_path(self, simple_rgba_image):
        """Test loading with pathlib.Path object."""
        # Should accept both str and Path objects
        pass
    
    def test_iterate_pixels(self, simple_rgba_image):
        """Test iterating over all pixels in row-major order."""
        # loader = ImageLoader(simple_rgba_image)
        # pixels = list(loader.iter_pixels())
        # assert len(pixels) == 9  # 3x3 image
        # First pixel should be red
        # assert pixels[0] == ((0, 0), (255, 0, 0, 255))
        pass
    
    def test_large_image_loading(self, large_image):
        """Test loading larger images (performance consideration)."""
        # loader = ImageLoader(large_image)
        # assert loader.width == 50
        # assert loader.height == 50
        pass
